requirements-min.txt
streamlit
fastembed
faiss-cpu
numpy
ollama
pypdf
